Ignore Krazy2 qclasses issues.
No Qt classes with a specific KDE subclass are really being used (except
in the tests). The objects may be stored as a pointer to a Qt class for
broader compatibility, but the objects are actually created from KDE
subclasses, or subclasses of those KDE subclasses.

The Qt classes are used in the tests, though, as they provide the
minimum API needed by the tested classes. There is no need for KDE
subclasses there.