Fix Krazy2 i18ncheckarg issues (ambiguous names).
The strings are not extracted from tests, and their context, besides the
role and subcue, does not affect the returned strings when using the
default translation (the one used in the tests). However, the context
was clarified anyway, as i18ncheckarg does not seem to support
"krazy:exclude" or "krazy:excludeall" markers.