Use the "standard" Messages.sh format.
The current Messages.sh files were a simplification of the Messages.sh
files used for applications outside KDE repositories. Instead of
simplifying the scripts it would have been better to replace them with
the "standard" Messages.sh format used all along the repository (of
course, with variations depending on the needs of each application).
This is what have been done now.

The change should also bring back the "NAME OF TRANSLATORS" and "EMAIL
OF TRANSLATORS" entries to the templates, missing since the
simplification of the scripts.

Besides those two extra entries (in the editor and the test application,
but not in the library), no other changes are expected, besides a
reorder of the entries in the template. That should not break any
translation, though, as the merge of templates and translations should
take care of the reorder. At least I hope so :)