Set proper copyright holder and year.
According to the source code of the Krazy2 copyright checker (the check
itself is not implemented yet), non-legal entities (like "KDE TechBase
Developers") are not allowed to hold copyright. The KTutorial Test
Application code is based on an old version of the code of the "Saving
and loading" tutorial from TechBase
(http://techbase.kde.org/Development/Tutorials/Saving_and_loading) that,
according to the page history, was created and modified only by Matt
Williams at that time (2007-12-27 14:08). Thus, the copyright holder is
set to Matt Williams and the copyright year is set to 2007.