Improve internationalization.
Before, the English name of the programming language was just appended
to the localized name of the tutorial. Besides not being
internationalized, appending the programming language name prevents a
proper localization in right-to-left languages. Now, the programming
language name and the "Tutorial name (Programming language name)" string
are both internationalized.