Clean a little the memory tests.
There is no need to look for "Veritas" in the stack, as it is the name
of a library used by the KDevelop plugin from which the memory tests
were got. For the same reason, there will not be any stack that contains
a file name with the "xtest" or "veritas" strings.