Remove new lines from debug output.
There is no need to add a new line at the end of several messages sent
to KDebug, as it is done automatically. Moreover, adding a new line
between messages in the same output should be avoided; either use two
different outputs, or a single output and everything on the same line.