Prevent tests from hang due to modal widgets.
Under some circumstances (like high system load) the target application
used in the tests may take more time than expected to start. In that
case, the RemoteObjectChooser may show a message box to inform the user
about the problem. If that happens, the test hangs until the message box
is closed.

A system to automatically close an unexpected message box was added. To
ensure that the message box is an unexpected one and thus is hanging the
test, it is closed once the timeouts used to close the expected message
boxes have expired.